ImageAnalysisMovingAverageAlongCL.py
This script takes the volume file of the lumen having the pixel values projected on, the centreline and surface of the lumen. It moves a sphere clipping filter along the centreline and calculates the average pixel value. Next, the average pixel value is projected into the centreline file as the array.